# ⚡ Performance improvements for /api/Leaderboard/Get