--- a/src/Application/Common/Interfaces/IDateTime.cs
+++ b/src/Application/Common/Interfaces/IDateTime.cs
@@ -2,4 +2,5 @@
 public interface IDateTime
 {
     DateTime Now { get; }
+    DateTime UtcNow { get; }
 }