--- a/src/Application/Leaderboard/Queries/GetLeaderboardList/GetLeaderboardListQuery.cs
+++ b/src/Application/Leaderboard/Queries/GetLeaderboardList/GetLeaderboardListQuery.cs
@@ -1,42 +1,75 @@
-﻿using MoreLinq;
-using SSW.Rewards.Application.Common.Extensions;
+﻿using System.Text.RegularExpressions;
 using SSW.Rewards.Shared.DTOs.Leaderboard;
 
 namespace SSW.Rewards.Application.Leaderboard.Queries.GetLeaderboardList;
 
-public class GetLeaderboardListQuery : IRequest<LeaderboardViewModel>
-{
-
-}
+public class GetLeaderboardListQuery : IRequest<LeaderboardViewModel> { }
 
 public class Handler : IRequestHandler<GetLeaderboardListQuery, LeaderboardViewModel>
 {
     private readonly IApplicationDbContext _context;
+    private readonly IDateTime _dateTime;
 
-    public Handler(IApplicationDbContext context)
+    public Handler(IApplicationDbContext context, IDateTime dateTime)
     {
         _context = context;
+        _dateTime = dateTime;
     }
 
     public async Task<LeaderboardViewModel> Handle(GetLeaderboardListQuery request, CancellationToken cancellationToken)
     {
+        DateTime utcNow = _dateTime.UtcNow;
+
         var users = await _context.Users
+            .AsNoTracking()
+            .AsSplitQuery()
             .TagWithContext()
-            .Where(u => u.Activated)
-            .Include(u => u.UserAchievements)
-                .ThenInclude(ua => ua.Achievement)
-            .Include(u => u.SocialMediaIds)
-                .ThenInclude(s => s.SocialMediaPlatform)
-            .Where(u => !string.IsNullOrWhiteSpace(u.FullName))
-            .Select(u => new LeaderboardUserDto(u))
+            .Where(x => x.Activated && !string.IsNullOrWhiteSpace(x.FullName))
+            .Select(x => new LeaderboardUserDto
+            {
+                UserId = x.Id,
+                Name = x.FullName,
+                Email = x.Email,
+                ProfilePic = x.Avatar,
+                PointsClaimed = x.UserRewards.Sum(ur => ur.Reward.Cost),
+                TotalPoints = x.UserAchievements.Sum(ua => ua.Achievement.Value),
+                PointsToday = x.UserAchievements
+                    .Where(ua =>
+                        ua.AwardedAt.Year == utcNow.Year &&
+                        ua.AwardedAt.Month == utcNow.Month &&
+                        ua.AwardedAt.Day == utcNow.Day)
+                    .Sum(ua => ua.Achievement.Value),
+                PointsThisWeek = x.UserAchievements
+                    .Where(ua => utcNow.AddDays(-7) <= ua.AwardedAt && ua.AwardedAt <= utcNow)
+                    .Sum(ua => ua.Achievement.Value),
+                PointsThisMonth = x.UserAchievements
+                    .Where(ua => ua.AwardedAt.Year == utcNow.Year && ua.AwardedAt.Month == utcNow.Month)
+                    .Sum(ua => ua.Achievement.Value),
+                PointsThisYear = x.UserAchievements
+                    .Where(ua => ua.AwardedAt.Year == utcNow.Year)
+                    .Sum(ua => ua.Achievement.Value),
+                Title = x.SocialMediaIds
+                    .Where(s => s.SocialMediaPlatform.Name == "Company")
+                    .Select(s => s.SocialMediaUserId)
+                    .FirstOrDefault()
+            })
             .ToListAsync(cancellationToken);
 
-        users
-            .OrderByDescending(lud => lud.TotalPoints)
-            .ForEach((u, i) => u.Rank = i + 1);
-        
-        var model = new LeaderboardViewModel { Users = users };
 
-        return model;
+        // Post-processing
+        int rank = 0;
+        Regex checkHttpRegex = new(@"^https?://", RegexOptions.Compiled, TimeSpan.FromMilliseconds(100));
+        foreach (LeaderboardUserDto? user in users.OrderByDescending(lud => lud.TotalPoints))
+        {
+            user.Rank = ++rank;
+            user.Title = user.Title switch
+            {
+                _ when !string.IsNullOrEmpty(user.Title) => checkHttpRegex.Replace(user.Title, string.Empty),
+                _ when user.Email?.EndsWith("ssw.com.au") == true => "SSW",
+                _ => "Community"
+            };
+        }
+
+        return new LeaderboardViewModel { Users = users };
     }
 }