--- a/src/Infrastructure/Services/DateTimeService.cs
+++ b/src/Infrastructure/Services/DateTimeService.cs
@@ -4,4 +4,5 @@ namespace SSW.Rewards.Infrastructure.Services;
 public class DateTimeService : IDateTime
 {
     public DateTime Now => DateTime.Now;
+    public DateTime UtcNow => DateTime.UtcNow;
 }