… src/Application/Leaderboard/Queries/GetFilteredLeaderboardList/GetFilteredLeaderboardList.cs
… src/Application/Leaderboard/Queries/GetFilteredLeaderboardList/GetFilteredLeaderboardListQuery.cs
… src/Application/Leaderboard/Queries/GetFilteredLeaderboardList/GetFilteredLeaderboardListValidator.cs
… src/Application/Leaderboard/Queries/GetFilteredLeaderboardList/GetFilteredListQueryValidator.cs