> 1. What triggered this change? (PBI link, Email Subject, conversation + reason, etc)

✏️ #1119 
This endpoint is called very frequently and can be very slow.

> 2. What was changed?

✏️ Rewrite query to be about 2x faster on my local machine.

When using 10 threads for 60 seconds it went from 6.8 req/sec to 15.5 req/sec. That's about 2.8 times faster under load.

Based on SQL Server execution plan analysis.

| Query | Memory grant | Memory Used | Estimated CPU Cost | CPU time |
| ------- | ----------------| ---------------- | ---------------------- | ---------- |
| Old     | 283 MB           | 8.5MB              | 0.954                         | 108 ms    |
| New   | 9.2 MB            | 9 MB                | 0.15                           | 48 ms      |

Also, fixed some incorrect calculations like balance.

Benchmark test:

```ps1
bombardier -c 10 -d 60s -m GET `
-H "Authorization: Bearer {{ TOKEN }}" `
https://localhost:5001/api/Leaderboard/Get
```

> 3. Did you do pair or mob programming?

✏️ No.
<!-- E.g. I worked with @gordonbeeming and @sethdailyssw -->

<!-- 
Check out the relevant rules
- https://www.ssw.com.au/rules/rules-to-better-pull-requests
- https://www.ssw.com.au/rules/write-a-good-pull-request
- https://www.ssw.com.au/rules/over-the-shoulder-prs 
- https://www.ssw.com.au/rules/do-you-use-co-creation-patterns
-->